refactor: move types under admin/config
