+++ src/admin/components/elements/Nav/minimal.ts
@@ -1,6 +1,6 @@
 import Drawer from '@mui/material/Drawer/Drawer.js';
 import { CSSObject, styled, Theme } from '@mui/material/styles/index.js';
-import { config } from '../../../../config/ui.js';
+import { config } from '../../../config/ui.js';
 
 const openedMixin = (theme: Theme): CSSObject => ({
   width: config.ui.navWidth,
