+++ src/admin/components/forms/fieldTypes/types.ts
@@ -1,5 +1,5 @@
 import { FieldValues, UseFormReturn } from 'react-hook-form';
-import { Field } from '../../../../config/types.js';
+import { Field } from '../../../config/types.js';
 
 export type Props = {
   form: UseFormReturn<FieldValues, any>;
