+++ src/admin/components/utilities/Auth/index.tsx
@@ -2,7 +2,7 @@ import React, { createContext, useCallback, useContext, useMemo, useState } from
 import { useNavigate } from 'react-router-dom';
 import useSWR from 'swr';
 import useSWRMutation from 'swr/mutation';
-import { AuthUser, Permission, PermissionsAction } from '../../../../config/types.js';
+import { AuthUser, Permission, PermissionsAction } from '../../../config/types.js';
 import { logger } from '../../../../utilities/logger.js';
 import { api, attachRetry, removeAuthorization, setAuthorization } from '../../../utilities/api.js';
 import { AuthContext } from './types.js';
