+++ src/admin/config/ui.ts
