+++ src/admin/pages/ContentType/Edit/CreateField/Context/types.ts
@@ -1,6 +1,6 @@
 import { SWRResponse } from 'swr';
 import { SWRMutationResponse } from 'swr/mutation';
-import { Collection, Field } from '../../../../../../config/types.js';
+import { Collection, Field } from '../../../../../config/types.js';
 
 export type FieldContext = {
   createField: () => SWRMutationResponse<Field, any, string, Record<string, any>>;
