+++ src/admin/pages/ContentType/Edit/EditField/Context/types.ts
@@ -1,5 +1,5 @@
 import { SWRMutationResponse } from 'swr/mutation';
-import { Field } from '../../../../../../config/types.js';
+import { Field } from '../../../../../config/types.js';
 
 export type FieldContext = {
   updateField: (id: number) => SWRMutationResponse<Field, any, string, Record<string, any>>;
