+++ src/admin/pages/Login/Context/index.tsx
@@ -1,6 +1,6 @@
 import React, { createContext, useContext, useMemo } from 'react';
 import useSWR, { SWRConfiguration } from 'swr';
-import { ProjectSetting } from '../../../../config/types.js';
+import { ProjectSetting } from '../../../config/types.js';
 import { api } from '../../../utilities/api.js';
 import { LoginContext } from './types.js';
