+++ src/admin/pages/Login/Context/types.ts
@@ -1,5 +1,5 @@
 import { SWRConfiguration, SWRResponse } from 'swr';
-import { ProjectSetting } from '../../../../config/types.js';
+import { ProjectSetting } from '../../../config/types.js';
 
 export type LoginContext = {
   getProjectSetting: (config?: SWRConfiguration) => SWRResponse<ProjectSetting>;
