+++ src/admin/pages/Project/Context/types.ts
@@ -1,6 +1,6 @@
 import { SWRConfiguration, SWRResponse } from 'swr';
 import { SWRMutationResponse } from 'swr/mutation';
-import { ProjectSetting } from '../../../../config/types.js';
+import { ProjectSetting } from '../../../config/types.js';
 
 export type ProjectSettingContext = {
   getProjectSetting: (config?: SWRConfiguration) => SWRResponse<ProjectSetting>;
