+++ src/admin/pages/Role/Context/types.ts
@@ -1,6 +1,6 @@
 import { SWRConfiguration, SWRResponse } from 'swr';
 import { SWRMutationResponse } from 'swr/mutation';
-import { Collection, Permission, Role } from '../../../../config/types.js';
+import { Collection, Permission, Role } from '../../../config/types.js';
 
 export type RoleContext = {
   getRoles: () => SWRResponse<Role[]>;
@@ -9,6 +9,8 @@ export type RoleContext = {
   updateRole: (id: string) => SWRMutationResponse<void, any, string, Record<string, any>>;
   getCollections: (config?: SWRConfiguration) => SWRResponse<Collection[]>;
   getPermissions: (id: string, config?: SWRConfiguration) => SWRResponse<Permission[]>;
-  createPermission: (id: string) => SWRMutationResponse<Permission, any, string, Record<string, any>>;
+  createPermission: (
+    id: string
+  ) => SWRMutationResponse<Permission, any, string, Record<string, any>>;
   deletePermission: (id: string, permissionId: string) => SWRMutationResponse;
 };
