+++ src/admin/pages/Role/Edit/HeaderCell/types.ts
@@ -1,4 +1,4 @@
-import { PermissionsAction } from '../../../../../config/types.js';
+import { PermissionsAction } from '../../../../config/types.js';
 
 export type Props = {
   action: PermissionsAction;
