+++ src/admin/pages/Role/Edit/index.tsx
@@ -23,7 +23,7 @@ import { Controller, SubmitHandler, useForm } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import { useNavigate, useParams } from 'react-router-dom';
 import { EmptyTable, MainCard } from 'superfast-ui';
-import { PermissionsAction, Role } from '../../../../config/types.js';
+import { PermissionsAction, Role } from '../../../config/types.js';
 import { logger } from '../../../../utilities/logger.js';
 import { DeleteButton } from '../../../components/elements/DeleteButton/index.js';
 import { Loading } from '../../../components/elements/Loading/index.js';
