+++ src/admin/pages/User/Context/types.ts
@@ -1,6 +1,6 @@
 import { SWRConfiguration, SWRResponse } from 'swr';
 import { SWRMutationResponse } from 'swr/mutation';
-import { Role, User } from '../../../../config/types.js';
+import { Role, User } from '../../../config/types.js';
 
 export type UserContext = {
   getUsers: () => SWRResponse<User[]>;
