+++ src/admin/pages/collections/Edit/index.tsx
@@ -6,7 +6,7 @@ import { useForm } from 'react-hook-form';
 import { useTranslation } from 'react-i18next';
 import { useNavigate, useParams } from 'react-router-dom';
 import { MainCard } from 'superfast-ui';
-import { Field } from '../../../../config/types.js';
+import { Field } from '../../../config/types.js';
 import { logger } from '../../../../utilities/logger.js';
 import { DeleteButton } from '../../../components/elements/DeleteButton/index.js';
 import { RenderFields } from '../../../components/forms/RenderFields/index.js';
