+++ src/admin/pages/collections/Edit/types.ts
@@ -1,4 +1,4 @@
-import { Collection } from '../../../../config/types.js';
+import { Collection } from '../../../config/types.js';
 
 export type Props = {
   collection: Collection;
